requirements-dev.txt
-r ./requirements-common.txt

caniusepython3==4.0.0
django-debug-toolbar==1.6
django-nose==1.4.4
nose==1.3.7
pep8==1.7.0
pylint==1.6.4
